+++ drf_jwt_2fa/urls.py
@@ -9,4 +9,5 @@ urlpatterns = [
     path("verify/", views.verify_auth_token, name="verify"),
     path("totp/setup/", views.setup_totp, name="totp-setup"),
     path("totp/confirm/", views.confirm_totp, name="totp-confirm"),
+    path("2fa-method/", views.set_2fa_method, name="set-2fa-method"),
 ]