+++ ChangeLog.rst
@@ -6,6 +6,8 @@ Unreleased
 
 * Add TOTP (Time-based One-Time Password) support
 
+* Add endpoint for changing the preferred 2FA method
+
 * Fire ``user_logged_in`` signal on successful authentication
 
   * Note: This updates ``user.last_login`` by default